pyproject.toml
[project]
name = "ea_assistant"
version = "0.1.0"
description = "GPT-powered local assistant system"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.9, <3.13"
dependencies = [
    "openai==1.12.0",
    "httpx==0.27.0",
    "pydantic==2.6.1",
    "qdrant-client==1.7.0",
    "sqlite-utils==3.35.2",
    "flask==3.0.2",
    "streamlit==1.31.1",
    "google-api-python-client==2.118.0",
    "google-auth-oauthlib==1.2.0",
    "requests==2.31.0",
    "python-frontmatter==1.0.0",
    "beautifulsoup4==4.12.3",
    "lxml==5.1.0",
    "numpy==1.26.4",
    "ruamel.yaml==0.18.6",
    "loguru==0.7.2",
    "structlog==24.1.0",
    "matplotlib==3.8.3",
    "pandas==2.2.1",
    "watchdog==3.0.0",
    "psutil==5.9.8",
]

[project.optional-dependencies]
dev = [
    "ruff==0.2.2",
    "black==24.2.0",
    "mypy==1.8.0",
    "pre-commit==3.6.0",
    "pytest==8.0.1",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "pytest-asyncio==0.23.5",
]
docs = [
    "mkdocs==1.5.3",
    "mkdocs-material==9.5.3",
    "mkdocstrings==0.24.0",
    "mkdocstrings-python==1.7.1",
    "mkdocs-autorefs==1.0.1",
    "mkdocs-jupyter==0.24.1",
]
ai = [
    "faiss-cpu==1.7.4",
    "torch==2.2.0",
    "transformers==4.37.2",
    "sentence-transformers==2.5.1",
]
test = [
    "pytest==8.0.1",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "pytest-asyncio==0.23.5",
    "pytest-xdist==3.5.0",
    "pytest-timeout==2.2.0",
    "pytest-env==1.1.3",
]

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "RUF"]
ignore = ["E402", "E741", "F841", "N802", "N803", "N806"]
fix = true

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src"
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
